Better Use for Hardware-Rendering + Readded Text-Output
* RGNDS::GL2D::PolyShape now uses the GPU to translate/move, rotate and scale its Points
* RGNDS::GL2D::PolyShape now requires a array of RGNDS::Point<double> instead of RGNDSD::Point<float>

+ added function RGNDS::GL2D::glText
  * can be give a Transform - Instance to define Position and Scale
  * positioning and scaling are done via GPU too

* made RGNDS::Engine::error output more immediat
* Definition of PI and PI2 is more accurate now

+++ gl2d_polyshape.h
@@ -31,7 +31,7 @@ namespace RGNDS {
 
             */
 
-            PolyShape(short numTris, const Point<float> aTris[], GL_GLBEGIN_ENUM rendermode);
+            PolyShape(short numTris, const Point<double> aTris[], GL_GLBEGIN_ENUM rendermode);
             virtual ~PolyShape();
 
             virtual void draw(unsigned short color, short screen=0, Transform* t = &Transform::_default);
@@ -43,7 +43,7 @@ namespace RGNDS {
 
         protected:
             short numPoints = 0;
-            Point<float> *points;
+            Point<double> *points;
             GL_GLBEGIN_ENUM rendertype;
 
         };
